pyproject.toml
[tool.poetry]
name = "mentoria-rag-corte"
version = "0.1.0"
description = "ETL/RAG pipelines with Ploomber"
authors = ["Your Name <you@example.com>"]
license = "Apache 2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
ploomber = "^0.23.0"
jupysql = "^0.10.1"
jupysql-plugin = "^0.2.2"
duckdb-engine = "^0.9.2"
ipykernel = "^6.25.2"
farm-haystack = {extras = ["colab", "faiss", "inference", "preprocessing"], version = "^1.21.2"}
python-dotenv = "^1.0.0"
openai = "^0.28.0"
ipywidgets = "^8.1.1"
chainlit = "^0.7.0"
pandas = "2.0.3"
striprtf = "^0.0.26"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
